\name{covEB-package}
\alias{covEB-package}
\docType{package}
\title{
\packageTitle{covEB}
}
\description{
\packageDescription{covEB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{covEB}
\packageIndices{covEB}
The function for this package is covEB that calculates an empirical Bayes estimate of a given covariance matrix assuming that is has a block diagonal structure.
}
\author{
\packageAuthor{covEB}

Maintainer: \packageMaintainer{covEB}
}

